Disable dysfunctional local storage feature.
It would seem local storage is no usable stand-in for a real DB. It gets filled up by other parties to the point where there's barely any space left for this script's data. Rendering it more or less useless after a few videos have been stored. Even if it is truncated the amount of videos it can hold is simply too small.